refactor!(php): remove phpactor.el
BREAKING CHANGE: This removes the phpactor.el package in favor of LSP
support via lsp-mode/eglot. Use `+lsp` with the `:lang php` module
instead (with one of the supported LSP servers).

This was done because phpactor.el is on the way out of maintainership,
is redundant with pre-existing phpactor support in lsp-mode/eglot (or
the other, possibly superior LSP servers), and to simplify modules for
whom LSP/Eglot is *the* way to get these features.